[ESKL-33/withdraw] - update dagger modules
