+++ app/src/main/java/com/asfoundation/wallet/withdraw/usecase/WithdrawFiatUseCase.kt
@@ -1,17 +1,17 @@
 package com.asfoundation.wallet.withdraw.usecase
 
-import cm.aptoide.skills.interfaces.EwtObtainer
+import com.asfoundation.wallet.ewt.EwtAuthenticatorService
 import com.asfoundation.wallet.withdraw.repository.WithdrawRepository
 import io.reactivex.Completable
 import java.math.BigDecimal
 
 class WithdrawFiatUseCase(
-  private val ewtObtainer: EwtObtainer,
+  private val ewtObtainer: EwtAuthenticatorService,
   private val repository: WithdrawRepository,
 ) {
 
   fun execute(email: String, amount: BigDecimal): Completable {
-    return ewtObtainer.getEWT()
+    return ewtObtainer.getEwtAuthentication()
       .flatMapCompletable { repository.withdrawAppcCredits(it, email, amount) }
   }
 }
